fix(backups): reloads repo-host when secrets are updated (#23)
Repo-host mounts the latest certificates as secrets upon creation. Every now and then
these certificates are updated (expiry, corruption, race condition, etc) and if repo-host is
not restarted, pgbackrest jobs will fail due to file mismatch.

This patch introduces an annotation that tracks pgbackrest secret `resourceVersion` in
repo-host, causing changes to the secret to trigger a repo-host restart and forcing it to
load the latest certificate file.

Signed-off-by: Juliana Oliveira <juliana@fly.io>